Removed runtime assertion: if a non-plain token is provided, the `SessionMiddleware` will discard it
This change comes from a mutation discovered due to missing `assert()` execution in test scenarios.

Specifically, the `assert()` was never negated via a mutation test: instead, we blindly accepted that
only `UnencryptedToken` can come from the parser (not declared in the parser interface).

Hereby, we verify that if something different from an `UnencryptedToken` is produced by an `Lcobucci\JWT\Parser`,
the `SessionMiddleware` will instead continue execution (no crash), and the token will be completely ignored
instead.

This is both better for performance and for reliability, when upgrading `lcobucci/jwt` in future, and removes
previously uncovered code mutations in `infection/infection`.